+++ gen_dartpy/example/lib/gen_dartpy_example.dart
@@ -1,5 +1,6 @@
 import 'package:dartpy/dartpy.dart';
 import 'package:dartpy/dartpy_annotations.dart';
+
 part 'gen_dartpy_example.g.dart';
 
 @PyFunction(module: 'multiply')
@@ -11,12 +12,16 @@ double mult_double(double a, double b) => pymultiplydouble(a, b);
 @PyFunction(module: 'multiply', name: 'multiply')
 num mult_num(num a, num b) => pymultiplynum(a, b);
 
+@PyFunction(module: 'os', name: 'getenv')
+String getenv(String envName) => pygetenv(envName);
+
 void main() {
   try {
     print(multiply(6, 4));
     print(mult_double(6.13, 5.2));
     print(mult_num(6, 5.2));
     print(mult_num(6, 2));
+    print(getenv('HOME'));
   } on DartPyException catch (e) {
     if (pyErrOccurred()) {
       dartpyc.PyErr_Print();