+++ gen_dartpy/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.0.2
+- String conversion fixes (used to convert to `bytes` object which is not the same as `str`)
+
+## 0.0.2
 - Update for python 3.12
 
 ## 0.0.1